Managers finished
Possibly change most of the managers later to "lazily" get information
that's not really needed for the task at hand, e.g. when calling
summaryManager.getSummariesById you don't really need the major,
department and school objects that come along with it (since they're
included in the course objects that you actually would want to get in
with that function call).